remove http-metamask-data-access-write
