+++ packages/request-client.js/src/http-metamask-data-access.ts
@@ -1,14 +1,15 @@
-import { CombinedDataAccess } from '@requestnetwork/data-access';
-import { ClientTypes, CurrencyTypes, DataAccessTypes } from '@requestnetwork/types';
+import { Block } from '@requestnetwork/data-access';
+import { requestHashSubmitterArtifact } from '@requestnetwork/smart-contracts';
+import { ClientTypes, CurrencyTypes, DataAccessTypes, StorageTypes } from '@requestnetwork/types';
 import { ethers } from 'ethers';
+import { EventEmitter } from 'events';
 import HttpDataAccess from './http-data-access';
-import { HttpDataAccessConfig, NodeConnectionConfig } from './http-data-access-config';
-import { HttpDataRead } from './http-data-read';
-import { HttpMetaMaskDataWrite } from './http-metamask-data-write';
+import { NodeConnectionConfig } from './http-data-access-config';
+
 /**
  * Exposes a Data-Access module over HTTP
  */
-export default class HttpMetaMaskDataAccess extends CombinedDataAccess {
+export default class HttpMetaMaskDataAccess extends HttpDataAccess {
   /**
    * Cache block persisted directly (in case the node did not have the time to retrieve it)
    * (public for easier testing)
@@ -21,9 +22,6 @@ export default class HttpMetaMaskDataAccess extends CombinedDataAccess {
 
   private provider: ethers.providers.JsonRpcProvider | ethers.providers.Web3Provider;
   private networkName: CurrencyTypes.EvmChainName = 'private';
-
-  private readonly dataAccessConfig: HttpDataAccessConfig;
-  private readonly dataAccess: HttpDataAccess;
 
   /**
    * Creates an instance of HttpDataAccess.
@@ -45,22 +43,14 @@ export default class HttpMetaMaskDataAccess extends CombinedDataAccess {
       httpConfig: {},
     },
   ) {
+    super({ httpConfig, nodeConnectionConfig });
+
     ethereumProviderUrl = ethereumProviderUrl ? ethereumProviderUrl : 'http://localhost:8545';
 
     // Creates a local or default provider
-    const provider = web3
+    this.provider = web3
       ? new ethers.providers.Web3Provider(web3)
       : new ethers.providers.JsonRpcProvider({ url: ethereumProviderUrl });
-
-    const dataAccessConfig = new HttpDataAccessConfig({ httpConfig, nodeConnectionConfig });
-    const reader = new HttpDataRead(dataAccessConfig);
-    const writer = new HttpMetaMaskDataWrite(dataAccessConfig, provider);
-
-    super(reader, writer);
-
-    this.dataAccessConfig = dataAccessConfig;
-    this.dataAccess = new HttpDataAccess(dataAccessConfig);
-    this.provider = provider;
   }
 
   /**
@@ -84,7 +74,104 @@ export default class HttpMetaMaskDataAccess extends CombinedDataAccess {
     channelId: string,
     topics?: string[],
   ): Promise<DataAccessTypes.IReturnPersistTransaction> {
-    return this.writer.persistTransaction(transactionData, channelId, topics);
+    if (!this.networkName) {
+      const network = await this.provider.getNetwork();
+
+      this.networkName =
+        network.chainId === 1 ? 'mainnet' : network.chainId === 4 ? 'rinkeby' : 'private';
+    }
+    const submitterContract = requestHashSubmitterArtifact.connect(
+      this.networkName,
+      this.provider.getSigner(),
+    );
+
+    // We don't use the node to persist the transaction, but we will Do it ourselves
+
+    // create a block and add the transaction in it
+    const block: DataAccessTypes.IBlock = Block.pushTransaction(
+      Block.createEmptyBlock(),
+      transactionData,
+      channelId,
+      topics,
+    );
+
+    // store the block on ipfs and get the the ipfs hash and size
+    const { ipfsHash, ipfsSize } = await this.dataAccessConfig.fetch<{
+      ipfsHash: string;
+      ipfsSize: number;
+    }>('POST', '/ipfsAdd', { data: block });
+
+    // get the fee required to submit the hash
+    const fee = await submitterContract.getFeesAmount(ipfsSize);
+
+    // submit the hash to ethereum
+    const tx = await submitterContract.submitHash(
+      ipfsHash,
+      /* eslint-disable no-magic-numbers */
+      ethers.utils.hexZeroPad(ethers.utils.hexlify(ipfsSize), 32),
+      { value: fee },
+    );
+
+    const ethBlock = await this.provider.getBlock(tx.blockNumber ?? -1);
+
+    // create the storage meta from the transaction receipt
+    const storageMeta: StorageTypes.IEthereumMetadata = {
+      blockConfirmation: tx.confirmations,
+      blockNumber: tx.blockNumber ?? -1,
+      blockTimestamp: ethBlock.timestamp,
+      fee: fee.toString(),
+      networkName: this.networkName,
+      smartContractAddress: tx.to ?? '',
+      transactionHash: tx.hash,
+    };
+
+    // Add the block to the cache
+    if (!this.cache[channelId]) {
+      this.cache[channelId] = {};
+    }
+    this.cache[channelId][ipfsHash] = { block, storageMeta };
+
+    const eventEmitter = new EventEmitter() as DataAccessTypes.PersistTransactionEmitter;
+    const result: DataAccessTypes.IReturnPersistTransactionRaw = {
+      meta: {
+        storageMeta: {
+          ethereum: storageMeta,
+          ipfs: { size: ipfsSize },
+          state: StorageTypes.ContentState.PENDING,
+          timestamp: storageMeta.blockTimestamp,
+        },
+        topics: topics || [],
+        transactionStorageLocation: ipfsHash,
+      },
+      result: {},
+    };
+
+    // When the ethereum transaction is mined, emit an event 'confirmed'
+    void tx.wait().then((txConfirmed) => {
+      // emit the event to tell the request transaction is confirmed
+      eventEmitter.emit('confirmed', {
+        meta: {
+          storageMeta: {
+            ethereum: {
+              blockConfirmation: txConfirmed.confirmations,
+              blockNumber: txConfirmed.blockNumber,
+              blockTimestamp: ethBlock.timestamp,
+              fee: fee.toString(),
+              networkName: this.networkName,
+              smartContractAddress: txConfirmed.to,
+              transactionHash: txConfirmed.transactionHash,
+            },
+            state: StorageTypes.ContentState.CONFIRMED,
+            timestamp: ethBlock.timestamp,
+          },
+          topics: topics || [],
+          transactionStorageLocation: ipfsHash,
+        },
+        result: {},
+      });
+    });
+
+    return Object.assign(eventEmitter, result);
   }
 
   /**
